requirements.txt
exa-py==1.0.12
Flask==3.0.3
groq==0.9.0
openai==1.35.10
flask-cors-4.0.1